requirements.txt
streamlit==1.28.0
plotly==5.18.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
fpdf2==2.7.5